Prefer type composition and reusing over common types
