+++ source/ArrowKeyStepper/index.js
@@ -1,4 +1,6 @@
 // @flow
 
+export type { ScrollIndices } from "./types";
+
 export { default } from "./ArrowKeyStepper";
 export { default as ArrowKeyStepper } from "./ArrowKeyStepper";
