+++ source/ArrowKeyStepper/types.js
@@ -1,0 +1,6 @@
+// @flow
+
+export type ScrollIndices = {
+  scrollToColumn: number,
+  scrollToRow: number
+};
