+++ source/AutoSizer/AutoSizer.js
@@ -1,5 +1,7 @@
 /** @flow */
 
+import type { Size } from "./types";
+
 import React from "react";
 import createDetectElementResize from "../vendor/detectElementResize";
 
@@ -8,11 +10,6 @@ import createDetectElementResize from "../vendor/detectElementResize";
  * Child component should not be declared as a child but should rather be specified by a `ChildComponent` property.
  * All other properties will be passed through to the child component.
  */
-
-type Size = {
-  height: number,
-  width: number
-};
 
 type Props = {
   /** Function responsible for rendering children.*/
