+++ source/AutoSizer/index.js
@@ -1,4 +1,6 @@
 // @flow
 
+export type { Size } from "./types";
+
 export { default } from "./AutoSizer";
 export { default as AutoSizer } from "./AutoSizer";
