+++ source/AutoSizer/types.js
@@ -1,0 +1,6 @@
+// @flow
+
+export type Size = {
+  height: number,
+  width: number
+};
