+++ source/Grid/index.js
@@ -1,5 +1,15 @@
 /** @flow */
 
+export type {
+  NoContentRenderer,
+  Alignment,
+  CellSize,
+  OverscanIndicesGetter,
+  RenderedSection,
+  CellRendererParams,
+  Scroll
+} from "./types";
+
 export { default } from "./Grid";
 export { default as Grid } from "./Grid";
 export {
