+++ source/List/List.js
@@ -1,15 +1,15 @@
 /** @flow */
 
 import type {
+  NoContentRenderer,
   Alignment,
-  CellSizeGetter,
-  OverscanIndicesGetterParams,
-  OverscanIndices,
+  CellSize,
+  OverscanIndicesGetter,
   RenderedSection,
   CellRendererParams,
-  RowRendererParams,
   Scroll as GridScroll
-} from "../types";
+} from "../Grid";
+import type { RowRenderer, RenderedRows, Scroll } from "./types";
 
 import Grid, { accessibilityOverscanIndicesGetter } from "../Grid";
 import React from "react";
@@ -24,80 +24,60 @@ import cn from "classnames";
  * This component renders a virtualized list of elements with either fixed or dynamic heights.
  */
 
-type RenderedRows = {
-  overscanStartIndex: number,
-  overscanStopIndex: number,
-  startIndex: number,
-  stopIndex: number
-};
-
-type Scroll = {
-  clientHeight: number,
-  scrollHeight: number,
-  scrollTop: number
-};
-
 type Props = {
   "aria-label"?: string,
 
   /**
-     * Removes fixed height from the scrollingContainer so that the total height
-     * of rows can stretch the window. Intended for use with WindowScroller
-     */
-  autoHeight?: boolean,
+   * Removes fixed height from the scrollingContainer so that the total height
+   * of rows can stretch the window. Intended for use with WindowScroller
+   */
+  autoHeight: boolean,
 
   /** Optional CSS class name */
   className?: string,
 
   /**
-     * Used to estimate the total height of a List before all of its rows have actually been measured.
-     * The estimated total height is adjusted as rows are rendered.
-     */
+   * Used to estimate the total height of a List before all of its rows have actually been measured.
+   * The estimated total height is adjusted as rows are rendered.
+   */
   estimatedRowSize: number,
 
   /** Height constraint for list (determines how many actual rows are rendered) */
   height: number,
 
   /** Optional renderer to be used in place of rows when rowCount is 0 */
-  noRowsRenderer: () => React.Element<*> | null,
+  noRowsRenderer: NoContentRenderer,
 
-  /**
-     * Callback invoked with information about the slice of rows that were just rendered.
-     */
+  /** Callback invoked with information about the slice of rows that were just rendered.  */
 
   onRowsRendered: (params: RenderedRows) => void,
 
   /**
-     * Callback invoked whenever the scroll offset changes within the inner scrollable region.
-     * This callback can be used to sync scrolling between lists, tables, or grids.
-     */
+   * Callback invoked whenever the scroll offset changes within the inner scrollable region.
+   * This callback can be used to sync scrolling between lists, tables, or grids.
+   */
   onScroll: (params: Scroll) => void,
 
   /** See Grid#overscanIndicesGetter */
-  overscanIndicesGetter: (
-    params: OverscanIndicesGetterParams
-  ) => OverscanIndices,
+  overscanIndicesGetter: OverscanIndicesGetter,
 
   /**
-     * Number of rows to render above/below the visible bounds of the list.
-     * These rows can help for smoother scrolling on touch devices.
-     */
+   * Number of rows to render above/below the visible bounds of the list.
+   * These rows can help for smoother scrolling on touch devices.
+   */
   overscanRowCount: number,
 
-  /**
-     * Either a fixed row height (number) or a function that returns the height of a row given its index.
-     * ({ index: number }): number
-     */
-  rowHeight: CellSizeGetter | number,
+  /** Either a fixed row height (number) or a function that returns the height of a row given its index.  */
+  rowHeight: CellSize,
 
   /** Responsible for rendering a row given an index; ({ index: number }): node */
-  rowRenderer: (params: RowRendererParams) => React.Element<*>,
+  rowRenderer: RowRenderer,
 
   /** Number of rows in list. */
   rowCount: number,
 
   /** See Grid#scrollToAlignment */
-  scrollToAlignment: "auto" | "end" | "start" | "center",
+  scrollToAlignment: Alignment,
 
   /** Row index to ensure visible (by forcefully scrolling if necessary) */
   scrollToIndex: number,
@@ -117,6 +97,7 @@ type Props = {
 
 export default class List extends React.PureComponent {
   static defaultProps = {
+    autoHeight: false,
     estimatedRowSize: 30,
     onScroll: () => {},
     noRowsRenderer: () => null,
