+++ source/List/index.js
@@ -1,4 +1,6 @@
 /** @flow */
 
+export type { RowRendererParams } from "./types";
+
 export { default } from "./List";
 export { default as List } from "./List";
