+++ source/List/types.js
@@ -1,0 +1,27 @@
+// @flow
+
+import React from "react";
+
+export type RowRendererParams = {
+  index: number,
+  isScrolling: boolean,
+  isVisible: boolean,
+  key: string,
+  parent: Object,
+  style: Object
+};
+
+export type RowRenderer = (params: RowRendererParams) => React.Element<*>;
+
+export type RenderedRows = {
+  overscanStartIndex: number,
+  overscanStopIndex: number,
+  startIndex: number,
+  stopIndex: number
+};
+
+export type Scroll = {
+  clientHeight: number,
+  scrollHeight: number,
+  scrollTop: number
+};
